warning fixes for plots directory
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@23664 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/plots/Boundary/QvisBoundaryPlotWindow.C
@@ -560,7 +560,7 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
     const stringVector &matNames = boundaryAtts->GetBoundaryNames();
     intVector selectedBoundaries;
     bool update = true;
-    int i;
+    size_t i;
     QListWidgetItem *item;
 
     multipleColorList->blockSignals(true);
@@ -568,7 +568,7 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
     //
     // Make a pass through the widget to see how many boundaries are selected.
     //
-    for(i = 0; i < multipleColorList->count(); ++i)
+    for(i = 0; i < (size_t)multipleColorList->count(); ++i)
     {
         if(multipleColorList->item(i)->isSelected())
             selectedBoundaries.push_back(i);
@@ -577,7 +577,7 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
     //
     // Modify the names.
     //
-    if(matNames.size() == multipleColorList->count())
+    if(matNames.size() == (size_t)multipleColorList->count())
     {
         // See if the strings are the same..
         bool same = true;
@@ -598,10 +598,10 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
         else
             update = false;
     }
-    else if(matNames.size() > multipleColorList->count())
+    else if(matNames.size() > (size_t)multipleColorList->count())
     {
         // Set all of the existing names.
-        for(i = 0; i < multipleColorList->count(); ++i)
+        for(i = 0; i < (size_t)multipleColorList->count(); ++i)
             UpdateItem(i);
 
         // Add new entries
@@ -620,7 +620,7 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
             UpdateItem(i);
 
         // Remove excess entries
-        int numEntries = multipleColorList->count();
+        size_t numEntries = multipleColorList->count();
         for(i = matNames.size(); i < numEntries; ++i)
         {
             QListWidgetItem *item = multipleColorList->takeItem(multipleColorList->count() - 1);
@@ -669,7 +669,7 @@ QvisBoundaryPlotWindow::UpdateMultipleArea()
     //
     if(matNames.size() > 0)
     {
-        int selectedIndex = (selectedBoundaries.size() > 0) ?
+        size_t selectedIndex = (selectedBoundaries.size() > 0) ?
             selectedBoundaries[0] : 0;
         // Make sure that the selected index is in the range of visible colors.
         if(selectedIndex >= matNames.size())
