+++ src/plots/WellBore/avtWellBoreNameMapper.C
@@ -165,7 +165,7 @@ void
 avtWellBoreNameMapper::SetScale(double s)
 {
     scale = s;
-    for (int i = 0; i < actors.size(); i++)
+    for (size_t i = 0; i < actors.size(); i++)
     {
         actors[i]->SetScale(s);
     }
@@ -192,7 +192,7 @@ avtWellBoreNameMapper::SetLabelColor(double col[3])
     labelColor[0] = col[0];
     labelColor[1] = col[1];
     labelColor[2] = col[2];
-    for (int i = 0; i < actors.size(); i++)
+    for (size_t i = 0; i < actors.size(); i++)
     {
         actors[i]->SetForegroundColor(labelColor);
     }
@@ -221,7 +221,7 @@ avtWellBoreNameMapper::SetLabelColor(double r, double g, double b)
     labelColor[0] = r; 
     labelColor[1] = g; 
     labelColor[2] = b; 
-    for (int i = 0; i < actors.size(); i++)
+    for (size_t i = 0; i < actors.size(); i++)
     {
         actors[i]->SetForegroundColor(labelColor);
     }
@@ -248,14 +248,14 @@ avtWellBoreNameMapper::SetLabelVisibility(bool labelsOn)
    labelVis = labelsOn;
    if (labelsOn)
    {
-       for (int i = 0; i < actors.size(); i++)
+       for (size_t i = 0; i < actors.size(); i++)
        {
            actors[i]->UnHide();
        }
    } 
    else 
    {
-       for (int i = 0; i < actors.size(); i++)
+       for (size_t i = 0; i < actors.size(); i++)
        {
            actors[i]->Hide();
        }
